refactor: move vault init code into seperate file
